+++ internal/dataplane/translator/translate_tlsroute_test.go
@@ -25,6 +25,7 @@ func TestIngressRulesFromTLSRoutesUsingExpressionRoutes(t *testing.T) {
 	testCases := []struct {
 		name                 string
 		tcpRoutes            []*gatewayapi.TLSRoute
+		services             []*corev1.Service
 		expectedKongServices []kongstate.Service
 		expectedKongRoutes   map[string][]kongstate.Route
 		expectedFailures     []failures.ResourceFailure
@@ -51,6 +52,23 @@ func TestIngressRulesFromTLSRoutesUsingExpressionRoutes(t *testing.T) {
 								},
 							},
 						},
+					},
+				},
+			},
+			// After https://github.com/Kong/kubernetes-ingress-controller/pull/5392
+			// is merged the backendRef will be checked for existence in the store
+			// so we need to add them here.
+			services: []*corev1.Service{
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Namespace: "default",
+						Name:      "service1",
+					},
+				},
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Namespace: "default",
+						Name:      "service2",
 					},
 				},
 			},
@@ -112,6 +130,35 @@ func TestIngressRulesFromTLSRoutesUsingExpressionRoutes(t *testing.T) {
 					},
 				},
 			},
+			// After https://github.com/Kong/kubernetes-ingress-controller/pull/5392
+			// is merged the backendRef will be checked for existence in the store
+			// so we need to add them here.
+			services: []*corev1.Service{
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Namespace: "default",
+						Name:      "service1",
+					},
+				},
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Namespace: "default",
+						Name:      "service2",
+					},
+				},
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Namespace: "default",
+						Name:      "service3",
+					},
+				},
+				{
+					ObjectMeta: metav1.ObjectMeta{
+						Namespace: "default",
+						Name:      "service4",
+					},
+				},
+			},
 			expectedKongServices: []kongstate.Service{
 				{
 					Service: kong.Service{
@@ -164,7 +211,10 @@ func TestIngressRulesFromTLSRoutesUsingExpressionRoutes(t *testing.T) {
 	for _, tc := range testCases {
 		tc := tc
 		t.Run(tc.name, func(t *testing.T) {
-			fakestore, err := store.NewFakeStore(store.FakeObjects{TLSRoutes: tc.tcpRoutes})
+			fakestore, err := store.NewFakeStore(store.FakeObjects{
+				TLSRoutes: tc.tcpRoutes,
+				Services:  tc.services,
+			})
 			require.NoError(t, err)
 			translator := mustNewTranslator(t, fakestore)
 			translator.featureFlags.ExpressionRoutes = true